Tests done, exceptions added, cors filter added

+++ src/main/java/facades/PersonFacade.java
@@ -12,6 +12,10 @@ import entities.CityInfo;
 import entities.Hobby;
 import entities.Person;
 import entities.Phone;
+import exceptions.CityNotFoundException;
+import exceptions.HobbyNotFoundException;
+import exceptions.MissingInputException;
+import exceptions.PersonNotFoundException;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -49,13 +53,22 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO addPerson(String firstName, String lastName, String email, String street, String zipCode, String city, List<PhoneDTO> phones, List<HobbyDTO> hobbies) {
+    public PersonDTO addPerson(String firstName, String lastName, String email, String street, String zipCode, String city, List<PhoneDTO> phones, List<HobbyDTO> hobbies) throws MissingInputException {
+        if (firstName.length() == 0 || lastName.length() == 0) {
+            throw new MissingInputException("First name and/or last name is missing");
+        }
+        if (email.length() == 0 || email.contains("@") == false) {
+            throw new MissingInputException("Email missing and/or does not contain @");
+        }
+        if (street.length() == 0 || zipCode.length() == 0 || zipCode.length() > 4 || city.length() == 0) {
+            throw new MissingInputException("Street, zipcode or city is missing");
+        }
+
         EntityManager em = emf.createEntityManager();
         try {
 
             Person p = new Person(firstName, lastName, email);
             Address a = new Address(street);
-            
 
             for (PhoneDTO phDTO : phones) {
                 Phone ph = new Phone(phDTO.getNumber(), phDTO.getDescription());
@@ -92,54 +105,71 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO deletePerson(int id) {
+    public PersonDTO deletePerson(int id) throws PersonNotFoundException {
         EntityManager em = emf.createEntityManager();
-        Person person = em.find(Person.class, id);
 
         try {
-            em.getTransaction().begin();
-            
-//            for(Hobby h : person.getHobbies()){
-//            person.removeHobby(h);
-//            }
-            for (Phone ph : person.getPhones()) {
-                em.remove(ph);
-            }
-            em.remove(person);
-            em.remove(person.getAddress());
+            Person person = em.find(Person.class, id);
+            if (person == null) {
+                throw new PersonNotFoundException("Could not delete. Person not found");
+            } else {
+                em.getTransaction().begin();
 
-            em.getTransaction().commit();
-            PersonDTO pDTO = new PersonDTO(person);
-            return pDTO;
+                for (Phone ph : person.getPhones()) {
+                    em.remove(ph);
+                }
+                em.remove(person);
+                em.remove(person.getAddress());
+
+                em.getTransaction().commit();
+                PersonDTO pDTO = new PersonDTO(person);
+                return pDTO;
+            }
+
         } finally {
             em.close();
         }
     }
 
     @Override
-    public PersonDTO editPerson(PersonDTO p) {
+    public PersonDTO editPerson(PersonDTO p) throws MissingInputException, PersonNotFoundException {
+        if (p.getFirstName().length() == 0 || p.getLastName().length() == 0) {
+            throw new MissingInputException("First name and/or last name is missing");
+        }
+        if (p.getEmail().length() == 0 || p.getEmail().contains("@") == false) {
+            throw new MissingInputException("Email missing and/or does not contain @");
+        }
+        if (p.getStreet().length() == 0 || p.getZipCode().length() == 0 || p.getZipCode().length() > 4 || p.getCity().length() == 0) {
+            throw new MissingInputException("Street, zipcode or city is missing");
+        }
+
         EntityManager em = emf.createEntityManager();
 
         try {
             Person person = em.find(Person.class, p.getId());
+            if (person == null) {
+                throw new PersonNotFoundException("Could not delete. Person not found");
+            } else {
 
-            person.setFirstName(p.getFirstName());
-            person.setLastName(p.getLastName());
-            person.setEmail(p.getEmail());
-            person.getAddress().setStreet(p.getStreet());
-            person.getAddress().getCityInfo().setZipcode(p.getZipCode());
-            person.getAddress().getCityInfo().setCity(p.getCity());
+                person.setFirstName(p.getFirstName());
+                person.setLastName(p.getLastName());
+                person.setEmail(p.getEmail());
+                person.getAddress().setStreet(p.getStreet());
+                person.getAddress().getCityInfo().setZipcode(p.getZipCode());
+                person.getAddress().getCityInfo().setCity(p.getCity());
 
-            for (HobbyDTO hDTO : p.getHobbies()) {
-                Hobby h = new Hobby(hDTO.getName(), hDTO.getDescription());
-                person.addHobby(h);
+                for (HobbyDTO hDTO : p.getHobbies()) {
+                    Hobby h = new Hobby(hDTO.getName(), hDTO.getDescription());
+                    person.addHobby(h);
+                }
+                for (PhoneDTO phDTO : p.getPhones()) {
+                    Phone ph = new Phone(phDTO.getNumber(), phDTO.getDescription());
+                    person.addPhone(ph);
+                }
+                PersonDTO pDTO = new PersonDTO(person);
+                return pDTO;
             }
-            for (PhoneDTO phDTO : p.getPhones()) {
-                Phone ph = new Phone(phDTO.getNumber(), phDTO.getDescription());
-                person.addPhone(ph);
-            }
-            PersonDTO pDTO = new PersonDTO(person);
-            return pDTO;
+
         } finally {
             em.close();
         }
@@ -147,14 +177,18 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO getPerson(String phoneNumber) {
+    public PersonDTO getPerson(String phoneNumber) throws PersonNotFoundException {
         EntityManager em = emf.createEntityManager();
         try {
             TypedQuery<Person> query
                     = em.createQuery("SELECT p FROM Person p JOIN p.phones ph WHERE ph.number = :phoneNumber", Person.class);
             query.setParameter("phoneNumber", phoneNumber);
             Person p = query.getSingleResult();
-            return new PersonDTO(p);
+            if (p == null) {
+                throw new PersonNotFoundException("Could not find person");
+            } else {
+                return new PersonDTO(p);
+            }
 
         } finally {
             em.close();
@@ -191,43 +225,58 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonsDTO getAllPersonsHobbies(String hobby) {
+    public PersonsDTO getAllPersonsHobbies(String hobby) throws HobbyNotFoundException {
         EntityManager em = emf.createEntityManager();
         try {
             TypedQuery<Person> query
                     = em.createQuery("SELECT p FROM Person p JOIN p.hobbies h WHERE h.name =:hobby", Person.class);
             query.setParameter("hobby", hobby);
             List<Person> personEntities = query.getResultList();
-            PersonsDTO all = new PersonsDTO(personEntities);
-            return all;
+            if (personEntities == null) {
+                throw new HobbyNotFoundException("Could not find any persons with the given hobby");
+            } else {
+                PersonsDTO all = new PersonsDTO(personEntities);
+                return all;
+            }
+
         } finally {
             em.close();
         }
     }
 
     @Override
-    public PersonsDTO getAllPersonsCity(String city) {
+    public PersonsDTO getAllPersonsCity(String city) throws CityNotFoundException {
         EntityManager em = emf.createEntityManager();
         try {
             TypedQuery<Person> query
                     = em.createQuery("SELECT p FROM Person p JOIN p.address a JOIN a.cityInfo ci WHERE ci.city =:city", Person.class);
             query.setParameter("city", city);
             List<Person> personEntities = query.getResultList();
-            PersonsDTO all = new PersonsDTO(personEntities);
-            return all;
+            if (personEntities == null) {
+                throw new CityNotFoundException("Could not find any persons living in the given city");
+            } else {
+                PersonsDTO all = new PersonsDTO(personEntities);
+                return all;
+            }
+
         } finally {
             em.close();
         }
     }
 
     @Override
-    public long getHobbyCount(String hobbyName) {
+    public long getHobbyCount(String hobbyName) throws HobbyNotFoundException {
         EntityManager em = emf.createEntityManager();
         try {
             Query personCount = em.createQuery("SELECT COUNT(p) FROM Person p JOIN p.hobbies h WHERE h.name=:hobbyName");
             personCount.setParameter("hobbyName", hobbyName);
             long result = (long) personCount.getSingleResult();
-            return result;
+            if (result == 0) {
+                throw new HobbyNotFoundException("There are no persons with the given hobby");
+            } else {
+                return result;
+            }
+
         } finally {
             em.close();
         }
